# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: NodePress REST API
  version: "1.0.0"
  description: |
    WordPress REST API v2-compatible endpoints for NodePress.
    Base path: /wp/v2 — matches WP REST API routing convention.
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin Bearer token. Set NODEPRESS_ADMIN_TOKEN env var.
        Development default: dev-admin-token.
  schemas:
...
    WpTerm:
      type: object
      description: |
        A taxonomy term (category or tag) following WP REST API v2 shape.
        Sprint 3 constraints: parent is always 0 (flat, no hierarchy).
        count is computed from term_relationships.
      required:
        - id
        - count
        - description
        - link
        - name
        - slug
        - taxonomy
        - parent
      properties:
        id:
          type: integer
          description: Unique identifier for the term.
          example: 1
        count:
          type: integer
          description: Number of published posts associated with this term.
          example: 3
        description:
          type: string
          description: HTML description of the term.
          example: ""
        link:
          type: string
          description: URL of the term archive.
          example: "/wp/v2/categories/1"
        name:
          type: string
          description: HTML title for the term.
          example: "Uncategorized"
        slug:
          type: string
          description: An alphanumeric identifier for the term unique to its type.
          example: "uncategorized"
        taxonomy:
          type: string
          description: Type attribution for the term (category or post_tag).
          enum:
            - category
            - post_tag
          example: "category"
        parent:
          type: integer
          description: |
            The parent term ID. Always 0 in Sprint 3 (flat taxonomy, no hierarchy).
          example: 0
    WpUserPublic:
      type: object
      description: |
        A user object returned by GET /wp/v2/users (public listing).
        Omits email and sensitive fields for public exposure.
        Mirrors WP REST API v2 user shape.
      required:
        - id
        - name
        - slug
        - url
        - description
        - link
        - avatar_urls
        - _links
      properties:
        id:
          type: integer
          description: Unique identifier for the user.
          example: 1
        name:
          type: string
          description: The name of the user (displayName or login fallback).
          example: "Admin User"
        slug:
          type: string
          description: URL-friendly username (maps to login).
          example: "admin"
        url:
          type: string
          description: Author archive URL (empty in NodePress v1).
          example: ""
        description:
          type: string
          description: Short biography (empty in NodePress v1).
          example: ""
        link:
          type: string
          description: Author archive link (empty in NodePress v1).
          example: ""
        avatar_urls:
          type: object
          description: Avatar URLs by size (empty in NodePress v1).
          additionalProperties:
            type: string
          example:
            "24": ""
            "48": ""
            "96": ""
        _links:
          type: object
          description: HAL links (empty in NodePress v1).
          additionalProperties: true
    WpUser:
      type: object
      description: |
        A user object returned by GET /wp/v2/users/me (authenticated context).
        Includes email and all fields available in the users table.
        Mirrors WP REST API v2 user shape with NodePress extensions.
      required:
        - id
        - name
        - slug
        - email
        - url
        - description
        - link
        - locale
        - nickname
        - registered_date
        - roles
        - capabilities
        - _nodepress
      properties:
        id:
          type: integer
          description: Unique identifier for the user.
          example: 1
        name:
          type: string
          description: The name of the user (displayName or login fallback).
          example: "Admin User"
        slug:
          type: string
          description: URL-friendly username (maps to login).
          example: "admin"
        email:
          type: string
          format: email
          description: The user's email address.
          example: "admin@example.com"
        url:
          type: string
          description: Author archive URL (empty in NodePress v1).
          example: ""
        description:
          type: string
          description: Short biography (empty in NodePress v1).
          example: ""
        link:
          type: string
          description: Author archive link (empty in NodePress v1).
          example: ""
        locale:
          type: string
          description: User locale (hardcoded to en_US in NodePress v1).
          example: "en_US"
        nickname:
          type: string
          description: User nickname (displayName or login fallback).
          example: "Admin User"
        registered_date:
          type: string
          format: date-time
          description: User registration date (createdAt from database).
          example: "2026-04-17T10:00:00.000Z"
        roles:
          type: array
          description: Array of role names (e.g., ["admin", "editor"]).
          items:
            type: string
          example:
            - "admin"
        capabilities:
          type: object
          description: Capability map for the user (e.g., edit_posts true).
          additionalProperties: true
          example:
            edit_posts: true
        _nodepress:
          type: object
          description: NodePress-specific user fields.
          required:
            - login
          properties:
            login:
              type: string
              description: Database login field (username).
              example: "admin"
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "UNAUTHORIZED"
        message:
          type: string
          example: "Valid Bearer token required"
        data:
          type: object
          properties:
            status:
              type: integer
              example: 401
paths:
  /wp/v2/users:
    get:
      operationId: listUsers
      summary: List users
      description: |
        Returns a list of users (public endpoint). Email and sensitive fields are omitted.
        Mirrors WP REST API GET /wp/v2/users behaviour for public consumption.
      tags:
        - Users
      parameters:
        - name: page
          in: query
          description: Current page of the collection.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Maximum number of items to return (max 100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: A list of users (public data only, no email).
          headers:
            X-WP-Total:
              description: Total number of users.
              schema:
                type: integer
            X-WP-TotalPages:
              description: Total number of pages.
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WpUserPublic"
  /wp/v2/users/me:
    get:
      operationId: getAuthenticatedUser
      summary: Retrieve authenticated user
      description: |
        Returns the currently authenticated user's full profile (including email).
        Requires a valid Bearer token. Anonymous requests receive 401.
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated user profile (with email).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WpUser"
        "401":
          description: Missing or invalid Bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wp/v2/categories:
    get:
      operationId: listCategories
      summary: List categories
      description: |
        Returns a list of category terms (taxonomy=category). Public endpoint.
        Mirrors WP REST API GET /wp/v2/categories behaviour.
      tags:
        - Taxonomies
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Limit results to those matching a string (name or slug).
          schema:
            type: string
        - name: order
          in: query
          description: Order terms ascending or descending.
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
        - name: orderby
          in: query
          description: Sort collection by term attribute.
          schema:
            type: string
            enum:
              - id
              - name
              - slug
              - count
            default: name
      responses:
        "200":
          description: A list of categories.
          headers:
            X-WP-Total:
              description: Total number of terms matching the query.
              schema:
                type: integer
            X-WP-TotalPages:
              description: Total number of pages.
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WpTerm"
  /wp/v2/categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique identifier for the term.
        schema:
          type: integer
    get:
      operationId: getCategory
      summary: Retrieve a category
      description: Retrieves a single category term by ID. Public endpoint.
      tags:
        - Taxonomies
      responses:
        "200":
          description: Category found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WpTerm"
        "404":
          description: Category not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wp/v2/tags:
    get:
      operationId: listTags
      summary: List tags
      description: |
        Returns a list of tag terms (taxonomy=post_tag). Public endpoint.
        Mirrors WP REST API GET /wp/v2/tags behaviour.
      tags:
        - Taxonomies
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Limit results to those matching a string (name or slug).
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
        - name: orderby
          in: query
          schema:
            type: string
            enum:
              - id
              - name
              - slug
              - count
            default: name
      responses:
        "200":
          description: A list of tags.
          headers:
            X-WP-Total:
              description: Total number of terms matching the query.
              schema:
                type: integer
            X-WP-TotalPages:
              description: Total number of pages.
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WpTerm"
  /wp/v2/tags/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique identifier for the term.
        schema:
          type: integer
    get:
      operationId: getTag
      summary: Retrieve a tag
      description: Retrieves a single tag term by ID. Public endpoint.
      tags:
        - Taxonomies
      responses:
        "200":
          description: Tag found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WpTerm"
        "404":
          description: Tag not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...